Added support for Promises in `onClick` prop
Often if a button triggers an asyncronous action, the handler will return a
Promise that is resolved if successful or rejected if unsuccessful. This
component can make use of this Promise to automatically update it's state
without relying on external state management through the use of refs, which
should cut down on boilerplate code in many cases.